separate expression suite and query suite
